feat: Add statusOptions to UserDashboard
The code changes in this commit add the `statusOptions` variable to the `UserDashboard` component. This variable is imported from the `../../data` module and is used to store the available options for filtering the user data. This addition enhances the functionality of the UserDashboard by providing more filtering options for the users.